requirements.txt
click
Jinja2
pyyaml
python_on_whales
graphviz
pydot
